RDMA/hns: Fix an cmd queue issue when resetting
mainline inclusion
from mainline-v5.8
commit 3ec5f54f7a0f
category: bugfix
bugzilla: NA
CVE: NA

The hns NIC driver divides the reset process into 3 status:
initialization, hardware resetting and softwaring restting. RoCE driver
gets reset status by interfaces provided by NIC driver and commands will
not be sent to the IMP if the driver is in any above status. The main
reason for this issue is that there is a time gap between status 1 and 2,
if the RoCE driver sends commands to the IMP during this gap, the IMP will
stop working because it is not ready.

Signed-off-by: Shunfeng Yang <yangshunfeng2@huawei.com>
Signed-off-by: Yangyang Li <liyangyang20@huawei.com>
Reviewed-by: chunzhi hu <huchunzhi@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>