RDMA/hns: Fix bug during CMDQ initialization
mainline inclusion
from mainline-v5.13
commit af06b628a6bd
category: bugfix
bugzilla: NA
CVE: NA

When re-insmod hns-roce-hw-v2.ko and init CMDQ, we need to reinit PI and CI of CMDQ.
But, after last rmmod hns-roce-hw-v2.ko, the CMDQ is NOT reset, firmware still is waiting software to send cmd..
So,  if we re-init PI first, the firmware would thought that the software sent new cmds. then it start to process CMDQ.
It may be a long time for driver. If driver start to init hardware and really send cmds, the status of PI and CI are unexpected.

Signed-off-by: Shunfeng Yang <yangshunfeng2@huawei.com>
Signed-off-by: Yangyang Li <liyangyang20@huawei.com>
Reviewed-by: chunzhi hu <huchunzhi@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>