RDMA/hns: Bugfix for posting multiple srq work request
mainline inclusion
from mainline-v5.3
commit 97545b10221
category: bugfix
bugzilla: NA
CVE: NA

When the user submits more than 32 work request to a srq queue
at a time, it needs to find the corresponding number of entries
in the bitmap in the idx queue. However, the original lookup
function named ffs only processes 32 bits of the array element,
When the number of srq wqe issued exceeds 32, the ffs will only
process the lower 32 bits of the elements, it will not be able
to get the correct wqe index for srq wqe.

Signed-off-by: Shunfeng Yang <yangshunfeng2@huawei.com>
Signed-off-by: Yangyang Li <liyangyang20@huawei.com>
Reviewed-by: chunzhi hu <huchunzhi@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>