net: hns3: change affinity_mask to numa node range
driver inclusion
category: bugfix
bugzilla: NA
CVE: NA

----------------------------

Currently, affinity_mask is set to a single value. As a result,
irqbalance becomes invalid in SUBSET or EXACT mode. To solve
this problem, change affinity_mask to a value range. In this
way, irqbalance can be performed on the CPU of the NUMA node.

Signed-off-by: Yufeng Mo <moyufeng@huawei.com>
Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Reviewed-by: Junxin Chen <chenjunxin1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>