net: hns3: replace snprintf with scnprintf in hns3_update_strings
mainline inclusion
from mainline-v5.6-rc1
commit bea5416561b1b997adbe3a04a5665e647c589a13
category: feature
bugzilla: NA
CVE: NA

----------------------------

snprintf returns the number of bytes that would be written, which may be
greater than the actual length to be written. Here use extra code to
handle this.

scnprintf returns the number of bytes that was actually written, just use
scnprintf to simplify the code.

Signed-off-by: Chen Zhou <chenzhou10@huawei.com>
Signed-off-by: David S. Milldder <davem@davemloft.net>
Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Reviewed-by: Junxin Chen <chenjunxin1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>