net: hns3: Fix for geneve tx checksum bug
mainline inclusion
from mainline-v5.9-rc4
commit a156998fc92d3859c8e820f1583f6d0541d643c3
category: bugfix
bugzilla: NA
CVE: NA

----------------------------

when skb->encapsulation is 0, skb->ip_summed is CHECKSUM_PARTIAL
and it is udp packet, which has a dest port as the IANA assigned.
the hardware is expected to do the checksum offload, but the
hardware will not do the checksum offload when udp dest port is
6081.

This patch fixes it by doing the checksum in software.

Reported-by: Li Bing <libing@winhong.com>
Signed-off-by: Yi Li <yili@winhong.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Reviewed-by: Junxin Chen <chenjunxin1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>