net: hns3: refactor dump mac tbl of debugfs
driver inclusion
category: feature
bugzilla: NA
CVE: NA

----------------------------

Currently, the debugfs command for ncl config is implemented by
"echo xxxx > cmd", and record the information in dmesg. It's
unnecessary and heavy. To improve it, create a single file
"mac_tbl" for it, and query it by command "cat mac_tbl", return
the result to userspace, rather than record in dmesg.

Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Reviewed-by: Junxin Chen <chenjunxin1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>