scsi: hisi_sas: use threaded irq to process CQ interrupts
mainline inclusion
from mainline-v5.10-rc4
commit 81f338e9709d
category: bugfix
bugzilla: NA
CVE: NA

Currently IRQ_EFFECTIVE_AFF_MASK is enabled for ARM_GIC and ARM_GIC3, so it
only allows a single target CPU in the affinity mask to process interrupts
and also interrupt thread, and the performance of using threaded irq is
almost the same as tasklet. But if the config is not enabled, the interrupt
thread will be allowed all the CPUs in the affinity mask. At that situation
it improves the performance (about 20%).

Note: IRQ_EFFECTIVE_AFF_MASK is configured differently for different
architecture chip, and it seems to be better to make it be configured
easily.

Link: https://lore.kernel.org/r/1579522957-4393-2-git-send-email-john.garry@huawei.com
Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Yang Xingui <yangxingui@huawei.com>
Reviewed-by: kangfenglong <kangfenglong@huawei.com>
Reviewed-by: ouyangdelong < ouyangdelong@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>