sw64: fix lockdep error
Sunway inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IB73UR

--------------------------------

Initialize a dynamically allocated struct bin_attribute so we
can make lockdep happy.  This is a new requirement for
attributes and initially this is only needed when lockdep is
enabled.  Lockdep gives a nice error when your attribute is
added to sysfs if you don't have this.

Signed-off-by: He Chuyue <hechuyue@wxiat.com>
Reviewed-by: He Sheng <hesheng@wxiat.com>
Signed-off-by: Gu Zitao <guzitao@wxiat.com>