sw64: fix a bug in huge_pte_offset()
Sunway inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IB73UR

--------------------------------

In the function huge_pte_offset(), when the value of pmd is detected as
"unpresent", it returns an unexpected "NULL" instead of the address of
the pmd.

This patch ensures that huge_pte_offset() returns the expected value.

Signed-off-by: Gao Chen <gaochen@wxiat.com>
Reviewed-by: He Sheng <hesheng@wxiat.com>
Signed-off-by: Gu Zitao <guzitao@wxiat.com>