migration: fix reference counting exception issue
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IAUHWX
CVE: NA

----------------------------------------------------------------------

In a live migration scenario, if the number of VFs at the
destination is greater than that at the source, the recovery
operation will fail and qemu will not complete the process
of closing the device FD after exiting.
Causes the device reference count to be abnormal, causing
the driver to be unable to be uninstalled.

Therefore, it is necessary to supplement the device exit
operation when the device exits.

Fixes: 3344c0cc132d ("hisi-acc-vfio-pci:add DFX for acc migration driver")
Signed-off-by: Longfang liu <liulongfang@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>